pyproject.toml
[tool.poetry]
name = "star-wars"
version = "0.1.0"
description = "Application to fetch all start wars planets and movies"
authors = ["Ankit Sagar <ankit@fyxee.com>"]
readme = "README.md"
packages = [{include = "star_wars"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "3.2"
psycopg2 = "^2.9.5"
pydantic = "^1.10.7"
django-stubs = "^1.16.0"
djangorestframework = "^3.14.0"
requests = "^2.28.2"
pyyaml = "^6.0"
django-extensions = "^3.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
